[change] printed LOC means now logical LOC.
Before LOC in printing means physical LOC,
but now it means logical LOC,
which is count of lines that have 1 or more tokens.

And files are no longer read to count LOC,
but LOC is cached from token list used to tokenize,
so efficiency is improved.

Line numbers of code are still physical.